# shared/api-contracts/openapi.yaml
openapi: 3.0.0

info:
  title: AutoBackup Manager API
  description: Unified backup management API for Android, Web, and Backend services
  version: 1.0.0
  contact:
    name: AutoBackup Team

servers:
  - url: http://localhost:8080/api
    description: Local development
  - url: https://api.autobackup.app/api
    description: Production

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      required:
        - id
        - email
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time

    Schedule:
      type: object
      required:
        - id
        - name
        - sourceUri
        - destinationUri
        - destinationType
        - triggerHour
        - triggerMinute
        - mode
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        sourceUri:
          type: string
        destinationUri:
          type: string
        destinationType:
          type: string
          enum: [GOOGLE_DRIVE, ONE_DRIVE, DROPBOX, S3, LOCAL]
        triggerHour:
          type: integer
          minimum: 0
          maximum: 23
        triggerMinute:
          type: integer
          minimum: 0
          maximum: 59
        mode:
          type: string
          enum: [COPY, MOVE]
        fileFilter:
          type: string
          nullable: true
        isEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    BackupRun:
      type: object
      required:
        - id
        - scheduleId
        - deviceId
        - startedAt
        - status
      properties:
        id:
          type: string
          format: uuid
        scheduleId:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [PENDING, RUNNING, SUCCESS, FAILED, PARTIAL]
        totalFiles:
          type: integer
        movedFiles:
          type: integer
        failedFiles:
          type: integer
        bytesCopied:
          type: integer
          format: int64
        errorMessage:
          type: string
          nullable: true

    FileLog:
      type: object
      required:
        - id
        - runId
        - fileName
        - fileSize
        - sourceUri
        - status
      properties:
        id:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        fileName:
          type: string
        fileSize:
          type: integer
          format: int64
        sourceUri:
          type: string
        destinationUri:
          type: string
          nullable: true
        status:
          type: string
          enum: [SUCCESS, FAILED, SKIPPED]
        errorCode:
          type: string
          nullable: true
        transferredAt:
          type: string
          format: date-time
          nullable: true

    Device:
      type: object
      required:
        - id
        - deviceName
        - lastSeenAt
      properties:
        id:
          type: string
          format: uuid
        deviceName:
          type: string
        lastSeenAt:
          type: string
          format: date-time

paths:
  /auth/register:
    post:
      summary: Register a new user account
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
        '409':
          description: Email already exists

  /auth/login:
    post:
      summary: Authenticate user and receive tokens
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  expiresIn:
                    type: integer
        '401':
          description: Invalid credentials

  /auth/refresh:
    post:
      summary: Refresh access token
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  expiresIn:
                    type: integer
        '401':
          description: Invalid refresh token

  /schedules:
    get:
      summary: List all backup schedules
      tags: [Schedules]
      security:
        - bearerAuth: []
      parameters:
        - name: deviceId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of schedules
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Schedule'
                  total:
                    type: integer
        '401':
          description: Unauthorized

    post:
      summary: Create a new backup schedule
      tags: [Schedules]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - sourceUri
                - destinationUri
                - destinationType
                - triggerHour
                - triggerMinute
                - mode
              properties:
                name:
                  type: string
                sourceUri:
                  type: string
                destinationUri:
                  type: string
                destinationType:
                  type: string
                  enum: [GOOGLE_DRIVE, ONE_DRIVE, DROPBOX, S3, LOCAL]
                triggerHour:
                  type: integer
                triggerMinute:
                  type: integer
                mode:
                  type: string
                  enum: [COPY, MOVE]
                fileFilter:
                  type: string
                isEnabled:
                  type: boolean
      responses:
        '201':
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'

  /schedules/{id}:
    get:
      summary: Get a specific schedule
      tags: [Schedules]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Schedule details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '404':
          description: Not found

    put:
      summary: Update a schedule
      tags: [Schedules]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Schedule updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'

    delete:
      summary: Delete a schedule
      tags: [Schedules]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Schedule deleted

  /backup-runs:
    get:
      summary: List backup runs
      tags: [Backup Runs]
      security:
        - bearerAuth: []
      parameters:
        - name: scheduleId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of backup runs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BackupRun'
                  total:
                    type: integer

  /backup-runs/{id}:
    get:
      summary: Get a specific backup run
      tags: [Backup Runs]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Backup run details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupRun'

  /backup-runs/{id}/files:
    get:
      summary: Get file transfer logs for a backup run
      tags: [Backup Runs]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: File logs for the run
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FileLog'
                  total:
                    type: integer

  /devices/register:
    post:
      summary: Register a new device
      tags: [Devices]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - deviceName
                - deviceToken
              properties:
                deviceName:
                  type: string
                deviceToken:
                  type: string
      responses:
        '201':
          description: Device registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'

  /devices:
    get:
      summary: List registered devices
      tags: [Devices]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of devices
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
                  total:
                    type: integer